scaling: Fix potential crash when scaling large images
Quick fix for now.
It would probably be better to change things so that the ImlibScaleInfo
arrays only contain the needed ranges.

https://git.enlightenment.org/old/legacy-imlib2/issues/26

+++ src/lib/scale.c
@@ -68,7 +68,7 @@ __imlib_CalcPoints(int sw, int dw_, int b1, int b2, bool aa, int up)
         else
         {
             for (i = 0; i < dd; i++)
-                p[b1 + i] = b1 + (i * ss) / dd;
+                p[b1 + i] = b1 + ((long long)i * ss) / dd;
             i = dw - b2;
         }
     }
